% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ATMQuote}
\alias{ATMQuote}
\title{At-the-money quote}
\usage{
ATMQuote(symbol, Exp)
}
\arguments{
\item{symbol}{yahoo stock symbol}

\item{Exp}{Expiration dates; YYYY-MM}
}
\value{
xts object
}
\description{
Returns quotes for at-the-money options on a given stock
}
\details{
Makes a call to getOptionChain and getQuote.  Returns quotes for options
struck at the price that is closest to the last quoted price of the
underlying. Does not support multiple expiration months.  ALPHA code.
}
\examples{
\dontrun{
ATMQuote('SPY')
}
}
\author{
Garrett See
}
\seealso{
ATM_k, getOptionChain, getQuote
}

