% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conId.twsContractDetails}
\alias{conId.twsContractDetails}
\title{twsContractDetails class conId extractor}
\usage{
\method{conId}{twsContractDetails}(x, ...)
}
\description{
twsContractDetails class conId extractor
}
\author{
Garrett See
}
\keyword{internal}

