% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eWrapper.data.BID_ASK_OHLCV}
\alias{eWrapper.data.BID_ASK_OHLCV}
\title{eWrapper for getting BID, ASK, LAST and OHLCV values from IB...}
\usage{
eWrapper.data.BID_ASK_OHLCV(n)
}
\arguments{
\item{n}{number of contracts being watched}
}
\value{
list of functions
}
\description{
eWrapper for getting BID, ASK, LAST and OHLCV values from IB
}
\details{
This is \code{eWrapper.data} modified to include the Open, High, Low, and
Close columns.
}
\author{
Samo Pahor
}
\seealso{
\code{eWrapper.data}
}

