% Generated by roxygen2 (4.0.2): do not edit by hand
\name{future_id}
\alias{future_id}
\title{Create future_series ids}
\usage{
future_id(root, month, year, format = "CYY", sep = "_")
}
\arguments{
\item{root}{character vector of names of contract roots}

\item{month}{numeric vector of months (3 letter abbreviations also work)}

\item{year}{numeric vector of years}

\item{format}{character string indicating how the id should be formatted.
See \code{\link[FinancialInstrument]{format_id}}}

\item{sep}{character that will separate root_id and suffix_id}
}
\value{
character vector of primary_ids for future_series instruments, OR,
if \code{root} is missing, a character vector of suffix_ids.
}
\description{
Create a vector of future_series ids.
}
\examples{
future_id(root=c("ES","YM"), month=c(3,6,9,12), year=2010:2011)
#getSymbols(future_id("VX",1:12,10:11),src='cfe')
future_id('VX',M2C(),10)
#future_id('VX',paste(M2C(),collapse=','),10)
}
\author{
Garrett See
}
\seealso{
\code{\link[FinancialInstrument]{format_id}} and
  \code{\link[FinancialInstrument]{parse_suffix}} for examples of id formats.
}

