\name{Mid}
\alias{Mid}
\alias{has.Mid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Return Mid column / check to see if data contains Mid column
}
\description{
\code{Mid} returns the Mid column of an xts object
\code{has.Mid} checks to see if an xts object has a Mid column
}
\usage{
Mid(x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
xts object
}
}
\value{
\code{Mid} returns a univariate xts object
\code{has.Mid} returns logical (boolean)
}
\author{
gsee
}
\note{
These should be exported from quantmod.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
Op, Hi, Lo, Cl, Bid, Ask,
has.Op, has.Cl
}

