% Generated by roxygen2 (4.0.2): do not edit by hand
\name{twsClock}
\alias{twsClock}
\title{A clock utility to display the TWS server time...}
\usage{
twsClock(format = "\%H:\%M:\%S", refresh = 1, verbose = TRUE)
}
\arguments{
\item{format}{format}

\item{refresh}{refresh rate}

\item{verbose}{be verbose about what clientId you are connected with?}
}
\value{
live updating current time as reported by \code{\link[IBrokers]{reqCurrentTime}}.
Press escape to exit clock utility.
}
\description{
A clock utility to display the TWS server time
}
\details{
adapted from the code in the links in references
}
\examples{
\dontrun{
twsClock()
twsClock("\%A \%d \%B \%Y \%I:\%M:\%OS3 \%p", 1e-3)
}
}
\references{
\url{http://stackoverflow.com/questions/5953718/overwrite-current-output-in-the-r-console}

\url{http://4dpiecharts.com/2011/05/11/a-clock-utility-via-console-hackery/}

\url{http://www.asciitable.com/}
}

